Put request data in post body instead of URL
Although we were correctly using the POST method instead of GET to
send API requests, we were incorrectly specifying the request data to
requests with `params` instead of `data` so it was being put in the
URL instead of the post body. This is problematic because the API
server limits the link of URLs much more than post bodies. This is now
fixed.