Clean up logging: use separate logger, use lazy interpolation
Instead of calling `logging.debug`, `logging.warning`, etc.,
instantiate a logger for the library and log through it instead, which
is the right way to log inside a library.

Use lazy string interpolation for log messages instead of f-string
interpolation, so interpolation doesn't have to happen unless messages
are actually logged. This is also considered a best practice, to the
extent that the newest pylint complains when you do it wrong.