Make TaskCache pass-through-compatible with session object
Add methods and functionality to the TaskCache object so that it can
be used as a drop-in replacement for the Toodledo session object. You
can now use a cache instead of accessing the API directly simply by
instantiating a TaskCache object and using it wherever you were
previously using the Toodledo session object.

The TaskCache object now does better error-checking to confirm that
you're not trying to use task fields that the library doesn't support
yet. Eventually I'll add that error checking to the uncached task
functions as well.

The TaskCache object now checks if you attempt to instantiate a cache
from a file that was previously instantiated with incompatible
parameters, e.g., if you specify `comp=0` now whereas before the cache
was created with `comp=None` or `comp=1`. Incompatible cache opens
will now raise an exception, which you can fix simply by deleting the
cache file so it gets recreated.

The unit tests have been enhanced to test both cached and uncached
access.