Revamp task cache for ease of use and better performance
The task cache now keeps tracks locally of changes made to tasks and
only fetches changes from the server when it needs to, which is when
you explicitly tell it to update or when you mark a repeating task
complete with `reschedule=1`.

The drop-in interoperability between the cache and session objects has
been improved by ensuring that all the same methods exist on both,
even when they're no-ops.

There's a new context-manager method on the cache object,
`caching_everything`, which you can use to temporarily override the
`comp` value specified when creating the cache so that tasks you add
or modify that don't match it are still preserved in the cache. When
the context exits, non-matching tasks are removed from the cache.

The `last_update` and `last_delete` properties have been removed from
the cache object.

The `AddTasks` and `EditTasks` functions now return the results
returned by the server as task objects.