Fix race condition in task cache when rescheduling a task
After rescheduling a task server-side, we sometimes need to go back to
the server to ask for the completed task that was just created so we
can add it to our cache. When we do that we need to ask for tasks
modified at least a second before our idea of the last edit timestamp
to avoid a race condition where that newly created task has the same
edit timestamp.