Fix several bugs in the task cache and add tests to cover them
When `GetTasks` is called with an ID, don't nest the returned task
inside an extra list.

In paranoid mode (used primarily by unit tests), don't try to check
fields that the caller didn't request from the server, since we don't
have them to check.

Fix a bug which caused the wrong tasks to sometimes be removed from
the cache after editing tasks, related to what value was specified to
`comp` when creating the task.

Fix a bug which was causing some task fields to be removed from the
cache incorrectly after editing tasks.