Make sure the date of dueTime in cache matches the date in dueDate
Although only the time component of the dueTime field is supposed to
matter, the date component is supposed to match what's in dueDate, so
update the caching code to keep them in sync.

Perhaps a better solution would be for our layer over the API to
return dueDime as a datetime.time object rather than a
datetime.datetime, but that's something we'll have to think about
implementing later since it's a backward-incompatible change.